refactor: rename network_types module to network
Per review: the module is now `hopr_types::network` (matching the `network`
feature name). Submodules: network::{types, session, errors}.
